# common: hypercombined bufferlist