--- a/src/common/buffer.cc
+++ b/src/common/buffer.cc
@@ -1085,7 +1085,7 @@ using namespace ceph;
     if (p == ls->end())
       seek(off);
     unsigned left = len;
-    for (std::list<ptr>::const_iterator i = otherl._buffers.begin();
+    for (buffers_t::const_iterator i = otherl._buffers.begin();
 	 i != otherl._buffers.end();
 	 ++i) {
       unsigned l = (*i).length();
@@ -1127,8 +1127,8 @@ using namespace ceph;
 
     // buffer-wise comparison
     if (true) {
-      std::list<ptr>::const_iterator a = _buffers.begin();
-      std::list<ptr>::const_iterator b = other._buffers.begin();
+      buffers_t::const_iterator a = _buffers.begin();
+      buffers_t::const_iterator b = other._buffers.begin();
       unsigned aoff = 0, boff = 0;
       while (a != _buffers.end()) {
 	unsigned len = a->length() - aoff;
@@ -1174,7 +1174,7 @@ using namespace ceph;
 
   bool buffer::list::is_aligned(unsigned align) const
   {
-    for (std::list<ptr>::const_iterator it = _buffers.begin();
+    for (buffers_t::const_iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it) 
       if (!it->is_aligned(align))
@@ -1184,7 +1184,7 @@ using namespace ceph;
 
   bool buffer::list::is_n_align_sized(unsigned align) const
   {
-    for (std::list<ptr>::const_iterator it = _buffers.begin();
+    for (buffers_t::const_iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it) 
       if (!it->is_n_align_sized(align))
@@ -1195,7 +1195,7 @@ using namespace ceph;
   bool buffer::list::is_aligned_size_and_memory(unsigned align_size,
 						  unsigned align_memory) const
   {
-    for (std::list<ptr>::const_iterator it = _buffers.begin();
+    for (buffers_t::const_iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it) {
       if (!it->is_aligned(align_memory) || !it->is_n_align_sized(align_size))
@@ -1205,7 +1205,7 @@ using namespace ceph;
   }
 
   bool buffer::list::is_zero() const {
-    for (std::list<ptr>::const_iterator it = _buffers.begin();
+    for (buffers_t::const_iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it) {
       if (!it->is_zero()) {
@@ -1217,7 +1217,7 @@ using namespace ceph;
 
   void buffer::list::zero()
   {
-    for (std::list<ptr>::iterator it = _buffers.begin();
+    for (buffers_t::iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it)
       it->zero();
@@ -1227,7 +1227,7 @@ using namespace ceph;
   {
     ceph_assert(o+l <= _len);
     unsigned p = 0;
-    for (std::list<ptr>::iterator it = _buffers.begin();
+    for (buffers_t::iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it) {
       if (p + it->length() > o) {
@@ -1341,7 +1341,7 @@ using namespace ceph;
   void buffer::list::rebuild(ptr& nb)
   {
     unsigned pos = 0;
-    for (std::list<ptr>::iterator it = _buffers.begin();
+    for (buffers_t::iterator it = _buffers.begin();
 	 it != _buffers.end();
 	 ++it) {
       nb.copy_in(pos, it->length(), it->c_str(), false);
@@ -1370,7 +1370,7 @@ using namespace ceph;
 	&& _len > (max_buffers * align_size)) {
       align_size = round_up_to(round_up_to(_len, max_buffers) / max_buffers, align_size);
     }
-    std::list<ptr>::iterator p = _buffers.begin();
+    buffers_t::iterator p = _buffers.begin();
     while (p != _buffers.end()) {
       // keep anything that's already align and sized aligned
       if (p->is_aligned(align_memory) && p->is_n_align_sized(align_size)) {
@@ -1395,7 +1395,7 @@ using namespace ceph;
         */
         offset += p->length();
         unaligned.push_back(*p);
-        _buffers.erase(p++);
+        p = _buffers.erase(p);
       } while (p != _buffers.end() &&
   	     (!p->is_aligned(align_memory) ||
   	      !p->is_n_align_sized(align_size) ||
@@ -1439,7 +1439,9 @@ using namespace ceph;
     _len += bl._len;
     if (!(flags & CLAIM_ALLOW_NONSHAREABLE))
       bl.make_shareable();
-    _buffers.splice(_buffers.end(), bl._buffers );
+    std::move(bl._buffers.begin(), bl._buffers.end(),
+      std::back_inserter(_buffers));
+    bl._buffers.clear();
     bl._len = 0;
     bl.last_p = bl.begin();
   }
@@ -1447,7 +1449,7 @@ using namespace ceph;
   void buffer::list::claim_append_piecewise(list& bl)
   {
     // steal the other guy's buffers
-    for (std::list<buffer::ptr>::const_iterator i = bl.buffers().begin();
+    for (buffers_t::const_iterator i = bl.buffers().begin();
         i != bl.buffers().end(); ++i) {
       append(*i, 0, i->length());
     }
@@ -1567,7 +1569,7 @@ using namespace ceph;
   void buffer::list::append(const list& bl)
   {
     _len += bl._len;
-    for (std::list<ptr>::const_iterator p = bl._buffers.begin();
+    for (buffers_t::const_iterator p = bl._buffers.begin();
 	 p != bl._buffers.end();
 	 ++p) 
       _buffers.push_back(*p);
@@ -1635,7 +1637,7 @@ using namespace ceph;
     if (n >= _len)
       throw end_of_buffer();
     
-    for (std::list<ptr>::const_iterator p = _buffers.begin();
+    for (buffers_t::const_iterator p = _buffers.begin();
 	 p != _buffers.end();
 	 ++p) {
       if (n >= p->length()) {
@@ -1655,7 +1657,7 @@ using namespace ceph;
     if (_buffers.empty())
       return 0;                         // no buffers
 
-    std::list<ptr>::const_iterator iter = _buffers.begin();
+    buffers_t::const_iterator iter = _buffers.begin();
     ++iter;
 
     if (iter != _buffers.end())
@@ -1666,7 +1668,7 @@ using namespace ceph;
   string buffer::list::to_str() const {
     string s;
     s.reserve(length());
-    for (std::list<ptr>::const_iterator p = _buffers.begin();
+    for (buffers_t::const_iterator p = _buffers.begin();
 	 p != _buffers.end();
 	 ++p) {
       if (p->length()) {
@@ -1684,7 +1686,7 @@ using namespace ceph;
     clear();
 
     // skip off
-    std::list<ptr>::const_iterator curbuf = other._buffers.begin();
+    buffers_t::const_iterator curbuf = other._buffers.begin();
     while (off > 0 &&
 	   off >= curbuf->length()) {
       // skip this buffer
@@ -1727,7 +1729,7 @@ using namespace ceph;
     //cout << "splice off " << off << " len " << len << " ... mylen = " << length() << std::endl;
       
     // skip off
-    std::list<ptr>::iterator curbuf = _buffers.begin();
+    buffers_t::iterator curbuf = _buffers.begin();
     while (off > 0) {
       ceph_assert(curbuf != _buffers.end());
       if (off >= (*curbuf).length()) {
@@ -1769,7 +1771,7 @@ using namespace ceph;
       if (claim_by) 
 	claim_by->append( *curbuf, off, howmuch );
       _len -= (*curbuf).length();
-      _buffers.erase( curbuf++ );
+      curbuf = _buffers.erase( curbuf );
       len -= howmuch;
       off = 0;
     }
@@ -1783,7 +1785,7 @@ using namespace ceph;
   {
     list s;
     s.substr_of(*this, off, len);
-    for (std::list<ptr>::const_iterator it = s._buffers.begin(); 
+    for (buffers_t::const_iterator it = s._buffers.begin(); 
 	 it != s._buffers.end(); 
 	 ++it)
       if (it->length())
@@ -1949,7 +1951,7 @@ int buffer::list::write_fd(int fd) const
   int iovlen = 0;
   ssize_t bytes = 0;
 
-  std::list<ptr>::const_iterator p = _buffers.begin();
+  buffers_t::const_iterator p = _buffers.begin();
   while (p != _buffers.end()) {
     if (p->length() > 0) {
       iov[iovlen].iov_base = (void *)p->c_str();
@@ -1998,7 +2000,7 @@ int buffer::list::write_fd(int fd, uint64_t offset) const
 {
   iovec iov[IOV_MAX];
 
-  std::list<ptr>::const_iterator p = _buffers.begin();
+  buffers_t::const_iterator p = _buffers.begin();
   uint64_t left_pbrs = _buffers.size();
   while (left_pbrs) {
     ssize_t bytes = 0;
@@ -2028,7 +2030,7 @@ __u32 buffer::list::crc32c(__u32 crc) const
   int cache_hits = 0;
   int cache_adjusts = 0;
 
-  for (std::list<ptr>::const_iterator it = _buffers.begin();
+  for (buffers_t::const_iterator it = _buffers.begin();
        it != _buffers.end();
        ++it) {
     if (it->length()) {
@@ -2075,7 +2077,7 @@ __u32 buffer::list::crc32c(__u32 crc) const
 
 void buffer::list::invalidate_crc()
 {
-  for (std::list<ptr>::const_iterator p = _buffers.begin(); p != _buffers.end(); ++p) {
+  for (buffers_t::const_iterator p = _buffers.begin(); p != _buffers.end(); ++p) {
     raw *r = p->get_raw();
     if (r) {
       r->invalidate_crc();
@@ -2102,7 +2104,7 @@ sha1_digest_t buffer::list::sha1()
  */
 void buffer::list::write_stream(std::ostream &out) const
 {
-  for (std::list<ptr>::const_iterator p = _buffers.begin(); p != _buffers.end(); ++p) {
+  for (buffers_t::const_iterator p = _buffers.begin(); p != _buffers.end(); ++p) {
     if (p->length() > 0) {
       out.write(p->c_str(), p->length());
     }
@@ -2219,7 +2221,7 @@ std::ostream& buffer::operator<<(std::ostream& out, const buffer::ptr& bp) {
 std::ostream& buffer::operator<<(std::ostream& out, const buffer::list& bl) {
   out << "buffer::list(len=" << bl.length() << "," << std::endl;
 
-  std::list<buffer::ptr>::const_iterator it = bl.buffers().begin();
+  buffer::list::buffers_t::const_iterator it = bl.buffers().begin();
   while (it != bl.buffers().end()) {
     out << "\t" << *it;
     if (++it == bl.buffers().end()) break;