… src/include/buffer.h
… src/include/buffer_raw.h
… src/kv/LevelDBStore.cc
… src/librbd/internal.cc
… src/messages/MDataPing.h
… src/msg/async/PosixStack.cc
… src/msg/async/rdma/RDMAConnectedSocketImpl.cc
… src/msg/simple/Pipe.cc
… src/os/ObjectStore.h
… src/os/bluestore/KernelDevice.cc
… src/os/filestore/FileJournal.cc
… src/rgw/rgw_file.cc
… src/test/bufferlist.cc
… src/test/rgw/test_rgw_compression.cc