Summary
=======
`bufferlist` offers `buffer::raw_combined` concept. It's basically a `buffer::raw` control structure allocated with its underlying data buffer in one go. This reduces pressure on memory allocator while improving memory locality. The pull requests extends the idea throughout following changes:
* `bufferlist::_buffers` container has been switched to ~~`boost::intrusive::list`~~  a plain, C-style list to gain control over memory management of its nodes.
* `buffer::ptr` has been combined with the intrusive hooks to form `ptr_node`. The extra benefit is avoidance of memcpy/moving in some cases. 
* `buffer::raw` got `std::aligned_storage` to hold one `ptr_node` that points to it. Of course, as these instances live on the same memory, their life-times must be controlled - `ptr` can never outlive `raw`. That's the exact reason why `ptr_node` is not re-assignable.

As a result data buffer, `buffer::raw`, `buffer::ptr` and the container's hook can be allocated together.

UPDATE: additionally the memory footprint of `buffer::ptr` has been minimized from 32 bytes to 24 thanks to the singly linked list.

Microbenchmarking
===============
For the 4 KB allocation size `bufferlist::append` typically work on, around 40% time reduction has been observed  (`1 - 0.226191 / 0.371490`).

Before: 1b57df1d514588b1c8232fe0c348b21fcc90c4ee
--------
```
$ ./bin/unittest_bufferlist --gtest_filter=BufferList.BenchAlloc
Note: Google Test filter = BufferList.BenchAlloc
[==========] Running 1 test from 1 test case.
[----------] Global test environment set-up.
[----------] 1 test from BufferList
[ RUN      ] BufferList.BenchAlloc
100000 alloc of size 32768 in 0.701351
100000 alloc of size 25000 in 0.678349
100000 alloc of size 16384 in 0.670223
100000 alloc of size 10000 in 0.643028
100000 alloc of size 8192 in 0.640630
100000 alloc of size 6000 in 0.421637
100000 alloc of size 4096 in 0.371490
100000 alloc of size 1024 in 0.357983
100000 alloc of size 256 in 0.328453
100000 alloc of size 32 in 0.320399
100000 alloc of size 4 in 0.321341
[       OK ] BufferList.BenchAlloc (5455 ms)
[----------] 1 test from BufferList (5455 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test case ran. (5455 ms total)
[  PASSED  ] 1 test.
```

After: 58e3a9ab470620a0f704e2c79908810b19e03c89
-----
```
$ ./bin/unittest_bufferlist --gtest_filter=BufferList.BenchAlloc
Note: Google Test filter = BufferList.BenchAlloc
[==========] Running 1 test from 1 test case.
[----------] Global test environment set-up.
[----------] 1 test from BufferList
[ RUN      ] BufferList.BenchAlloc
100000 alloc of size 32768 in 0.614499
100000 alloc of size 25000 in 0.598030
100000 alloc of size 16384 in 0.589918
100000 alloc of size 10000 in 0.557164
100000 alloc of size 8192 in 0.547145
100000 alloc of size 6000 in 0.222730
100000 alloc of size 4096 in 0.226191
100000 alloc of size 1024 in 0.225212
100000 alloc of size 256 in 0.225188
100000 alloc of size 32 in 0.223975
100000 alloc of size 4 in 0.217744
[       OK ] BufferList.BenchAlloc (4248 ms)
[----------] 1 test from BufferList (4248 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test case ran. (4248 ms total)
[  PASSED  ] 1 test.
```