@mattbenjamin 2018-11-01
@rzarzynski this work is great;  we've wanted to do this for years, not clear how we never arrived all the way, I'm delighted;  rename "hangable_ptr" please, though!