@rzarzynski 2018-11-04
jenkins retest this please (`run-make-check.sh Terminated because the system killed it. Surely a premature die of the VM` on x86)