@tchaikov 2018-12-02
```
<error>
  <unique>0xbf</unique>
  <tid>1</tid>
  <kind>Leak_PossiblyLost</kind>
...
    <frame>
      <ip>0xA809D61</ip>
      <obj>/usr/lib64/valgrind/vgpreload_memcheck-amd64-linux.so</obj>
      <fn>posix_memalign</fn>
      <dir>/builddir/build/BUILD/valgrind-3.13.0/coregrind/m_replacemalloc</dir>
      <file>vg_replace_malloc.c</file>
      <line>1020</line>
    </frame>
    <frame>
      <ip>0xE8557E</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>ceph::buffer::list::append_hole(unsigned int)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/common</dir>
      <file>buffer.cc</file>
      <line>206</line>
    </frame>
    <frame>
      <ip>0x6F32C4</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>pool_pg_num_history_t::encode(ceph::buffer::list&amp;) const</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/osd</dir>
      <file>osd_types.h</file>
      <line>5684</line>
    </frame>
    <frame>
      <ip>0x69937D</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>OSD::handle_osd_map(MOSDMap*)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/osd</dir>
      <file>osd_types.h</file>
      <line>5733</line>
    </frame>
    <frame>
      <ip>0x6A8978</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>OSD::_dispatch(Message*)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/osd</dir>
      <file>OSD.cc</file>
      <line>6979</line>
    </frame>
    <frame>
      <ip>0x6A8D34</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>OSD::ms_dispatch(Message*)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/osd</dir>
      <file>OSD.cc</file>
      <line>6682</line>
    </frame>
...
```

```
    <frame>
      <ip>0xE8557E</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>ceph::buffer::list::append_hole(unsigned int)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/common</dir>
      <file>buffer.cc</file>
      <line>206</line>
    </frame>
    <frame>
      <ip>0xBC7BCD</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>FileJournal::prepare_entry(std::vector&lt;ObjectStore::Transaction, std::allocator&lt;ObjectStore::Transaction&gt; &gt;&amp;, ceph::buffer::list*)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/os</dir>
      <file>ObjectStore.h</file>
      <line>1414</line>
    </frame>
    <frame>
      <ip>0xA23D25</ip>
      <obj>/usr/bin/ceph-osd</obj>
      <fn>FileStore::queue_transactions(boost::intrusive_ptr&lt;ObjectStore::CollectionImpl&gt;&amp;, std::vector&lt;ObjectStore::Transaction, std::allocator&lt;ObjectStore::Transaction&gt; &gt;&amp;, boost::intrusive_ptr&lt;TrackedOp&gt;, ThreadPool::TPHandle*)</fn>
      <dir>/usr/src/debug/ceph-14.0.1-1370-gd1d4355/src/os/filestore</dir>
      <file>FileStore.cc</file>
      <line>2307</line>
    </frame>

...
```

and more 

see http://pulpito.ceph.com/kchai-2018-12-02_07:20:25-rados-wip-kefu2-testing-2018-12-02-0912-distro-basic-smithi/3302785/