@tchaikov 2018-12-05
http://pulpito.ceph.com/kchai-2018-12-05_13:56:41-rados-wip-kefu2-testing-2018-12-05-2007-distro-basic-smithi/